user authorization for changes

+++ controllers/courses.js
@@ -46,12 +46,22 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
   if (!bootcamp) {
     return next(
       new errorResponse(`No bootcamp with id ${req.params.bootcampId}!`, 404)
+    );
+  }
+
+  if (bootcamp.user.toString() != req.user.id && req.user.role != "admin") {
+    return next(
+      new errorResponse(
+        `User with id ${req.user.id} is not authorized to add a course ${bootcamp._id}`,
+        401
+      )
     );
   }
 
@@ -74,6 +84,15 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     return next(new errorResponse(`No course with id ${req.params.id}!`, 404));
   }
 
+  if (course.user.toString() != req.user.id && req.user.role != "admin") {
+    return next(
+      new errorResponse(
+        `User with id ${req.user.id} is not authorized to update a course ${bootcamp._id}`,
+        401
+      )
+    );
+  }
+
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -94,6 +113,15 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     return next(new errorResponse(`No course with id ${req.params.id}!`, 404));
+  }
+
+  if (course.user.toString() != req.user.id && req.user.role != "admin") {
+    return next(
+      new errorResponse(
+        `User with id ${req.user.id} is not authorized to delete a course ${bootcamp._id}`,
+        401
+      )
+    );
   }
 
   await course.deleteOne();
